chore: Update BASE_URL constant and fetchTeam function in Home component
Update the BASE_URL constant in the constant.js file to use the 'http://192.168.1.49:8000/' URL instead of 'http://192.168.1.215:8000/'. Additionally, modify the fetchTeam function in the Home component to fetch team data and update the 'team' state. These changes ensure that the correct API endpoint is used and the team data is fetched correctly.