refactor: Remove unused Assign component and update Task component
The Assign component was deleted as it was no longer being used. Additionally, the Task component was updated to fetch task details and handle task actions such as start, pause, resume, end, and add comments. The fetchTask function was moved outside of the useEffect hook to ensure it is called whenever there is a change in the tasks state. The handleAddAssigne function was also updated to pause the task before assigning it to a user. The fetchTask function is now called after each task action to update the task details.