refactor: Add react-select dependency and update TeamView component
This commit adds the react-select dependency to the project and updates the TeamView component in the ManageTeam directory. The react-select library is used for creating select dropdowns in the UI. The TeamView component now imports the Select component from the react-select library instead of the custom Select component from the project's index file.

Refactor the code to use the new dependency and component.